requirements.txt
ipykernel
unstructured
tiktoken
pinecone-client
PyPDF2
openai
langchain
pandas 
numpy
python-dotenv
streamlit
langchain_google_genai
google-generativeai
chromadb
faiss-cpu
